Installed Jest and Puppeteer Deleted src/App.test.js so it doesn't get run by jest.
(from previous commit)
Started adding tests for the bezKoder app.
See README.md for details.

Known issues:
If a test fails, Jest throws an error about not closing something
related to async().  I tried a few things to fix this but it still
happens.

Need to add better logging when a test fails.